refactor: Slim down main() funtion/file
Start segregating k8s operator/controller logic into its own module for
ease of manipulating codebase.
