+++ src/operator/endpoint_slice.rs
